rag-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.47.1
PyPDF2==3.0.1
google-generativeai==0.8.5
python-dotenv==1.1.1
python-multipart==0.0.6
PyJWT==2.8.0